Extract JSON methods to module
